Add signing in, board, and fix email verification
This commit adds the ability to sign in, and a board that contains an array of colors/pixels. Additionally, the email verification process has been fixed to not send multiple emails. This was due to the email verification process occuring on render of the page, which React-Strict-Mode does twice. The email verification process now only occurs when the user clicks the signUp button after correctly filling out the form.